+++ packages/flowr_dart/lib/src/base.dart
@@ -4,8 +4,11 @@ import 'package:meta/meta.dart' show mustCallSuper;
 
 import 'mixin.dart';
 
-/// 基础 flow
-abstract class BaseFlowR<T> with DisposeMx {
+@Deprecated('use FlowrMx')
+abstract class BaseFlowR<T> extends IService with FlowRMx<T> {}
+
+/// flowr mixin
+mixin FlowRMx<T> on IService {
   /// put new value
   T put(T value);
 
@@ -17,7 +20,10 @@ abstract class BaseFlowR<T> with DisposeMx {
 
   /// get current value
   T get value;
+}
 
+/// service
+abstract class IService with DisposeMx {
   @mustCallSuper
   @override
   dispose() {}