+++ packages/flowr_dart/lib/src/flowr.dart
@@ -7,6 +7,10 @@ import 'package:meta/meta.dart'
 
 import 'package:rxdart/rxdart.dart';
 
+/// FrService
+abstract class FrService extends IService
+    with LoggableMx, SlowlyMx, RunCatchingMx, SubsAutoDisposeMx {}
+
 /// FlowR
 /// --- Basic mixin ---
 /// [BaseFlowR] 核心基础功能: 使用Stream传递数据
@@ -19,8 +23,7 @@ import 'package:rxdart/rxdart.dart';
 /// 注意:
 /// - 不要在[FlowR]内部存储任何状态数据:
 ///   而应该在[T]value中存储, [tag] 代表[T]value(Model)的实例, 而非[FlowR] (ViewModel)的实例
-abstract class FlowR<T> extends BaseFlowR<T>
-    with LoggableMx, SlowlyMx, RunCatchingMx, UpdatableMx, SubsAutoDisposeMx {
+abstract class FlowR<T> extends FrService with FlowRMx<T>, UpdatableMx {
   /// set [put] log type
   LogExtra? get logExtra => LogExtra.self;