+++ packages/flowr_arch/CHANGELOG.md
@@ -1,4 +1,10 @@
+## 0.4.0 2025-10-29
+refactor adp flowr 2.3.0 FlowRMx, FrService
+fix FrRepo::stream
+feat sembast_test.dart
+
 ## 0.3.0 2025-8-31
+* feat IRepo::stream, FrRepo::stream
 
 ## 0.2.0 2025-8-27
 * feat